// example/auth/client/src/vo_gen.ts
/* eslint:disable */
// Code generated by gotsrpc https://github.com/foomo/gotsrpc/v2  - DO NOT EDIT.
import * as github_com_foomo_gotsrpc_v2_example_auth_service from './vo_gen'; // ./client/src/vo_gen.ts to ./client/src/vo_gen.ts

// github.com/foomo/gotsrpc/v2/example/auth/service.HelloRequest
export interface HelloRequest {
	message:string;
}
// github.com/foomo/gotsrpc/v2/example/auth/service.HelloResponse
export interface HelloResponse {
	message:string;
	user:github_com_foomo_gotsrpc_v2_example_auth_service.User;
}
// github.com/foomo/gotsrpc/v2/example/auth/service.LoginRequest
export interface LoginRequest {
	username:string;
	password:string;
}
// github.com/foomo/gotsrpc/v2/example/auth/service.LoginResponse
export interface LoginResponse {
	token:string;
	user:github_com_foomo_gotsrpc_v2_example_auth_service.User;
}
// github.com/foomo/gotsrpc/v2/example/auth/service.User
export interface User {
	id:string;
	username:string;
	email:string;
	loginAt:number;
}
// end of common js